fix(policy-as-code): Move and rename utility modules to 'utils' package
