+++ agents/policy-as-code/docs/LOW_LEVEL_DETAILS.md
@@ -77,7 +77,7 @@ To enhance efficiency and provide a consistent user experience, the agent includ
 **C. `configurations/core_policies` Document**
 *   `policies` (Array[String]): List of active "Core Policy" queries.
 
-### 4. LLM & Code Generation (`policy_as_code/llm_utils.py`)
+### 4. LLM & Code Generation (`policy_as_code/utils/llm.py`)
 
 The agent uses a sophisticated prompt engineering approach to generate accurate Python code.
